Fix content preventing MkDocs site from building
The MkDocs site that I'm [working on][1] would not build before this
change. After this change, the MkDocs site builds. This strange
behavior seems to be due to a [bug][2] with pymdown-extensions which
I reported.

I'm committing this to `master` instead of to my migration branch
because I'm trying to handle as much content clean-up in `master` as
possible so as to narrow the window during which I'll be resolving
merge conflicts.

[1]: https://github.com/centerofci/mathesar/issues/3079
[2]: https://github.com/facelessuser/pymdown-extensions/issues/2148